convert iD.Detect() iife to a module
