--- .eslintignore
@@ -1,2 +1,0 @@
-js/id/start.js
-js/id/end.js
