--- js/id/end.js
@@ -1,1 +1,0 @@
-})();
