--- js/id/start.js
@@ -1,1 +1,0 @@
-(function () {
